requirements.txt
requests
pandas
numpy
google-cloud-storage
google-cloud-bigquery
pyarrow
statsmodels
scikit-learn
python-dateutil